app/requirements/common.txt
Django==1.10.1
Pillow==3.3.0
psycopg2==2.6.2
djangorestframework==3.4.6
django-filter==0.14.0
python-dateutil==2.5.3
pytz==2016.6.1
django-timezone-field==1.3.0
celery==3.1.17
django-extensions==1.7.3
dj-database-url==0.4.1
django-markdownx==1.6.3